fix: include circular reference
