+++ src/tplot/TradePlot.mts
@@ -1,12 +1,7 @@
 import * as z from 'zod/v4'
-import {
-  type IAssetQuoteResponse,
-  type IQuoteBarEma,
-  type ITicker,
-  zTicker,
-} from '../models/ticker-info.mjs'
 import type { ICreatedBy, IUpdatedBy } from '../models/id-created-updated.mjs'
 import { type ISubplot, Subplot } from './Subplot.mjs'
+import { type ITicker, zTicker } from '../models/ticker-info.mjs'
 import {
   PriceInDollars,
   getAsNumber,
@@ -15,8 +10,6 @@ import {
 } from '../primitives/number-helper.mjs'
 import { zDateTime, zStringMinMax } from '../services/zod-helper.mjs'
 import { type IIdRequired } from '../models/IdManager.mjs'
-import { type IIdValue } from '../models/IdValueManager.mjs'
-import { type ITradePlotProfitizer } from './TradePlotProfitizer.mjs'
 import { isNullOrUndefined } from '../primitives/object-helper.mjs'
 import { newGuid } from '../primitives/uuid-helper.mjs'
 import { safeArray } from '../primitives/array-helper.mjs'
@@ -342,11 +335,4 @@ export class TradePlot implements ITradePlot {
       ? (this.getTargetHigh() - price) / price
       : undefined
   }
-}
-
-export type ITradePlotProfitizerWithContext = {
-  emas: IIdValue<ISubplot['id'], IQuoteBarEma[]>[]
-  profitizer?: ITradePlotProfitizer
-  quote?: IAssetQuoteResponse
-  tradePlot: ITradePlot
 }
